Revert "Feat: 플레이리스트 상세 UI 작업 완료 및 관련 파일 정리"
This reverts commit b16d1d4e17beace2e2a1e1cd42844884fa779295.
