+++ src/pages/DesignSystem/DesignSystem.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react'
 import BackButton from '@/shared/components/BackButton/BackButton'
 import BottomNav from '@/shared/components/BottomNav/BottomNav'
 import LikeIcon from '@/shared/components/stats/LikeIcon'
@@ -10,7 +9,8 @@ import HashTag from '@/shared/components/HashTag/HashTag'
 import MoreMenu, { MenuItem } from '@/shared/components/MoreMenu/MoreMenu'
 import StatusButton from '@/shared/components/StatusButton/StatusButton'
 import VideoPlayer from '@/shared/components/VideoPlayer/VideoPlayer'
-import CommentInput from '@/features/CommentInput/components/CommentInput'
+import CommentBox from '@/features/CommentBox/components/CommentBox'
+import { useState } from 'react'
 
 export default function DesignSystem() {
   const tabs = [
@@ -216,7 +216,7 @@ export default function DesignSystem() {
       </section>
 
       <div>
-        <CommentInput />
+        <CommentBox />
       </div>
     </div>
   )
