+++ src/shared/components/ui/button.tsx
@@ -25,7 +25,7 @@ const buttonVariants = cva(
       },
     },
     defaultVariants: {
-      variant: 'ghost',
+      variant: 'default',
       size: 'default',
     },
   },
